idutils: Restructure module to be configurable and readable
* Adds an entrypoint to register new custom custom schemes
* Ensures that the new schemes are not overriding existing ones
* Adds invenio-app as test dependency required by pytest-invenio
