+++ setup.cfg
@@ -37,6 +37,7 @@ install_requires =
 
 [options.extras_require]
 tests =
+    invenio-app>=1.4.0
     pytest-black-ng>=0.4.0
     pytest-cache>=1.0
     pytest-runner>=2.6.2
@@ -44,6 +45,16 @@ tests =
     sphinx>=4.5
 # Kept for backwards compatibility
 docs =
+
+[options.entry_points]
+invenio_base.apps =
+    idutils = idutils:IDUtils
+invenio_base.api_apps =
+    idutils = idutils:IDUtils
+invenio_base.finalize_app =
+    idutils = idutils.ext:finalize_app
+invenio_base.api_finalize_app =
+    idutils = idutils.ext:api_finalize_app
 
 [build_sphinx]
 source-dir = docs/
