+++ tests/mock_module/filters.py
@@ -1,0 +1,22 @@
+# -*- coding: utf-8 -*-
+#
+# This file is part of IDUtils
+# Copyright (C) 2024 CERN.
+#
+# IDUtils is free software; you can redistribute it and/or modify
+# it under the terms of the Revised BSD License; see LICENSE file for
+# more details.
+#
+# In applying this license, CERN does not waive the privileges and immunities
+# granted to it by virtue of its status as an Intergovernmental Organization
+# or submit itself to any jurisdiction.
+
+"""Mock module custom scheme filters."""
+
+
+def custom_scheme_filter(value):
+    """Define filter for `custom_scheme`.
+
+    Exclude `orcid` scheme.
+    """
+    return ["orcid"]
